use boost::numeric cast to convert int128_t to std::intmax_t for certain cases
- in eval_node.hpp, the factorial lambda returns converted type
- in stcc algorithm, size of Eigen::Matrix has to be int type

+++ examples/stcc/stcc.cpp
@@ -141,7 +141,8 @@ ExprPtr biorthogonal_transform(
   {
     using namespace Eigen;
     // Dimension of permutation matrix is n_particles!
-    const auto n = sequant::factorial(n_particles);
+    const auto n =
+        boost::numeric_cast<std::intmax_t>(sequant::factorial(n_particles));
 
     // Permutation matrix
     Eigen::Matrix<double, Dynamic, Dynamic> M(n, n);
